Fix identification of control-plane Node in e2e test framework (#3740)
We use the `node-role.kubernetes.io/<ROLE>` label to identify the
control-plane Node. The `<ROLE>` we use needs to depend on the K8s
version because:
- before K8s v1.20, the label is `node-role.kubernetes.io/master`
- starting with K8s v1.24, the label is `node-role.kubernetes.io/control-plane`.
- in between, both labels are used by K8s

However, the function used to determine the correct label to use based
on the K8s version was called *before* the K8s server version was
actually determined, so it was always returning
`node-role.kubernetes.io/master`. This was causing tests to fail for K8s
v1.24 (which has just been released), since the label no longer exists
in that version.

Signed-off-by: Antonin Bas <abas@vmware.com>