Support rendering changelog for patch releases
Update the script to support rendering changelog for patch releases.
The idea is to list PRs with the "kind/cherry-pick" label and find
the original PRs. Then retrieve the information of the original PR if
it contains the label "action/release-note".

This takes more time as we cannot check the PRs in one shot.

Signed-off-by: Xu Liu <xliu2@vmware.com>