Stop using k8s.io/component-base/logs to configure logging
K8s is deprecating most of the klog files that we use in Antrea (e.g.,
--log_file). To avoid any disruption (and to avoid the deprecation
warnings which are now displayed when running an Antrea component), we
stop using k8s.io/component-base/logs and instead configure klog
directly.

There is no significant change in the code base, as klog is still used
as the logging facade. We use klog as the logging backend (same as
before), with the possibility to move to something else in the future.

Fixes #3786

Signed-off-by: Antonin Bas <abas@vmware.com>