[ExternalNode] Add InterfaceConfig type for ExternalEntity (#3768)
As an ExternalEntity is generate according to one Interface defined
in an ExternalNode, one InterfaceConfig is created for the ExternalEnity
accordingly.
For ExternalNode scenario, Antrea Agent should connect the host network
interface to OVS as the uplink, and create a new host internal port to
take the uplink's network configurations. An IntefaceConfig for
ExteranlEntity uses the name of the host internal port, and maintains
the OpenFlow ports of the OVS port pair.

Signed-off-by: wenyingd <wenyingd@vmware.com>