Fix error loop caused by failure to delete a non-existing route (#3827)
When a LoadBalancer with ingress IP is created, corresponding route
for the ingress IP is also installed. When deleting the LoadBalancer,
corresponding route should be also uninstalled. Before uninstalling
the route, if the route is removed, then deleting the route will
get a failure and return an error. The error will trigger LoadBalancer
deletion again, and obviously, deleting the route will get a failure
and return an error another time, resulting in an error loop. This PR
fixes error loop by printing log instead of returning error when getting
the error "no such process" (such error means that the route expected to
be deleted doesn't exist).

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>