Fix Egress not working with kube-proxy IPVS strictARP mode
Check the arp_ignore sysctl value of the transport interface
and start a userspace ARP responder if it has a value other
than 0.
Copy the assigned IPs on the dummy interface to the ARP/NDP
responder on initializing to fix an issue that the responders
may not work as expected after the agent restarts.

Fixes: #3804

Signed-off-by: Xu Liu <xliu2@vmware.com>