Add indirect cycle handling to Grafana Sankey panel (#3820)
Sankey panel is rendered by Google Charts. Google Charts will not
render if a cycle exists. It can be a direct cycle: a->a (e.g., intra-Node
traffic in node_to_node_flows dashboard) or an indirect cycle: a->b,
b->c, c->a. Add an extra space to all the destination names
to avoid introducing cycles.

Signed-off-by: heanlan <hanlan@vmware.com>