[ExternalNode]Create/Delete AntreaAgentInfo based on ExternalNode (#3757)
Antrea controller watches ExternalNode create and delete event.
It creates AntreaAgentInfo whose name is the same as ExternalNode name
after ExternalNode is created and delete AntreaAgentInfo when
ExternalNode is deleted.

The change also refactors monitoring part for Node case and lets controller
create AntreaAgentInfo for Node too. With this change,
for both Node and ExternalNode cases, it is always
Antrea controller to create/delete AntreaAgentInfo and
it is always Antrea agent to update AntreaAgentInfo.

Signed-off-by: Mengdie Song <songm@vmware.com>