Fix ARP storming in TrafficControl redirect
When a target port receives an ARP request, if the ARP request is sent back to
OVS via a return port, then the ARP request will be flooded again and the
target port will receive the ARP request again. As a result, the ARP request
will be flooded again and again. This PR fixes the issue by is by dropping ARP
packets from return ports.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>