[ExternalNode] Fix antctl e2e test
Command "antctl proxy --agent-node" requires the information from AntreaAgentInfo.
As a result, we need to make sure the content of AntreaAgentInfo is populated before
we run the command.

Since we recently move AntreaAgentInfo creation from agent to controller, agent may
not set the content for AntreaAgentInfo when it starts if controller has not created the AntreaAgentInfo.
In this case, it will take a minute for agent to update the content in its next try.

This change adds retry logics in the antctl proxy e2e test and it will only invoke the command
after AntreaAgentInfo is ready.

Fixes:#3856

Signed-off-by: Mengdie Song <songm@vmware.com>