implement network policy for multicast (#3660)
For multicast traffic, we support ingress rules for IGMP, and egress
rules for multicast data traffic. And apply NetworkPolicy to real
traffic for both. Ingress for multicast traffic is not supported now.

While egress for IGMP only supports IGMP report, which is handled by
packetIn. This patch maintains a rule map for each group address to
fetch the rule which matches the member and also has the highest
priority. And packetIn will decide to allow or drop the IGMP report
traffic based on the matched rule.

Signed-off-by: Bin Liu <biliu@vmware.com>