[ExternalNode] Agent datapath implementation (#3769)
* [ExternalNode] Agent datapath implementation

1. Set up OF pipeline for ExternalNode
2. Install Openflow entries for ExternalNode connectivity
3. Install Openflow entries for ANP on ExternalNode case
4. Supporing installing Openflow entries to bypass traffic to/from
   special peer address on a protocol
5. Unify the flows to process the packets with ct_state match in
   networkpolicy feature between K8s cluster and external node cases

* Fix e2e test issue

The original e2e test cases only check IngressRuleTable/EgressTable flow
count, and use the result to decide the if the NP rule is realized. But
these two tables are used to realize K8s NP rules. For a normal ANP
rule, AntreaPolicyIngressRuleTable/AntreaPolicyEgressRuleTable are used,
but never checked in the case. Besides, the original case always uses
number 2 to check the rule flow existence, but the fact is there always
no less than 2 flows in IngressRule/EgressRule table, which are used to
bypass the established and related packets in a valid connection. So the
case actually doesn't check if the rule is realized or not.

Signed-off-by: wenyingd <wenyingd@vmware.com>