Improve support for Egress in Traceflow
Add an Observation in result of Traceflow for Egress which
include Egress component, two TraceflowAction, Egress IP
and Egress name.

Add a mark `RemoteSNATRegMark` to indicate remote SNAT for
remote Egress. This mark helps to distinguish this traffic from
Inter-Node Pod-to-Pod traffic on source Node when Egress IP is
equal to a remote Node IP (static Egress).

Add e2e test for Egress in Traceflow for two cases.
1. Egress from local Node.
2. Egress from remote Node.

Modify compareObservations to match Egress and EgressIP.

Add unit tests for this patch.

Fixes #2247

Signed-off-by: Kumar Atish <atish.iaf@gmail.com>