Support dynamic ofport on OVS gateway/tunnel port (#3623)
Originally the OpenFlow entries are using static port numbers for
gateway/tunnel/uplink which are used in the ofport_request fields in OVS
port creation. This is possible to introduce traffic issues if the
actual numbers assigned by OVS are different from the requests.

With this change, the predefined port numbers are still used in the OVS
port creation request, and save the actual port numbers in NodeConfig
after the OVS ports are created. And for the OpenFlow entries, the
actual port numbers are used.

Signed-off-by: wenyingd <wenyingd@vmware.com>