Refactor Pod creation for e2e tests
Functions createPodOnNode and CreatePodOnNodeInNamespace have long lists
of parameters, and some of them are almost never used. Their naming is
also confusing since they both let you specify an arbitrary Namespace,
and the only difference is that CreatePodOnNodeInNamespace takes a
container name as parameter, while createPodOnNode auto-generates it
based on the container image name.

We try a different approach, using a "PodBuilder" object. The advantage
is the ability to add new methods to the builder, with explicit names,
instead of adding new function parameters or using the mutate function.

Signed-off-by: Antonin Bas <abas@vmware.com>