Release workflow: trigger website update after assets upload
When the antrea.io website is updated after an Antrea Github release, it
is in charge of updating the Helm repo index. This operation will not
succeed if the Helm archive doesn't exist yet. The archive is uploaded
as a release asset by a different workflow from the one triggering the
website update. Both of these workflows are triggered by the same event
type (release creation) and if they run concurrently, the website update
will happen before release assets are uploaded, and the update will
therefore fail.

To avoid this issue, we unify the 2 workflows in one "process release"
workflow, with one job to upload assets (including the Helm archive) and
one job to trigger the website update. The second job depends on the
first one, and will not run until that one completes.

Signed-off-by: Antonin Bas <abas@vmware.com>