Fix FlowExporter memory bloat when export process is dead (#3994)
When flow exporter is enabled, but failed to connect to downstream IPFIX
collector, connections added to the priority queue inside flow exporter
won't be expired and removed from queue, causing memory to
bloat. Furthermore, connection store polling will remove conn from its
connections map when the flow is no longer in conntrack, but not the
related items in priority queue. When a new flow with same flow key is
reestablished, a duplicated item with same key will be added to the
queue, while the reference to the old one is lost, essentially causing
memory leak.

This change addresses above issue in the following aspects:

* Connection store polling removes stale conn from both connections map
  and the priority queue. Since CS polling is independent of exporting
  process liveness, this allows clean up to be done without connection
  to collector.
* Fixes init of Connection.LastExportTime to be connection start time to
  make sure CS polling logic works properly when the exporting process
  is dead. Previously LastExportTime will only be filled by exporting
  process at the time of export, causing zero value to be compare in
  certain cases.
* Adds guards in priority queue to prevent having two Connections with
  same connection key in the heap.

Benchmark test BenchmarkExportConntrackConns did not show observable
difference before and after change.

Fixes item 1 and 2 in #3972. Severity of item 3 is lower, which will be
addressed in a later change.

Signed-off-by: Shawn Wang <wshaoquan@vmware.com>