Fix broken Markdown links check for main branch (#4018)
When check-modified-files-only is specified, base-branch will be used as
the branch to compare to find modified files. It defaults to "master"
which doesn't exist in Antrea. It should be set to the target branch of
the pull request. After it's corrected, there is no difference between
main and release branches. Therefore, we could just merge them to one
job.

Signed-off-by: Quan Tian <qtian@vmware.com>