Remove dependency on github.com/satori/go.uuid module
Most of our code uses github.com/google/uuid, but the NetworkPolicy
controller was using github.com/satori/go.uuid.
github.com/satori/go.uuid has a security issue (CVE-2021-3538) which was
fixed in the master branch but is not part of any released version. It
seems that the package is not actively maintained any more.
Even though this security issue doesn't impact us, it can still show up
in some security scans. Therefore, we remove the dependency and instead
use github.com/google/uuid consistently throughout the code. Plus, one
less dependency is always nice :)

The new UUIDs are the same as the old ones, which is not really a strict
requirement given that on Controller restart, all the information is
computed again and sent to all Agents.

Signed-off-by: Antonin Bas <abas@vmware.com>