Handle Node restarts better in EKS
We try to limit the probability of Pods being created without the Antrea
CNI being invoked. For example:

1) a Node is restarted
2) the aws-node Pod is restarted and overwrite the CNI conf file (i.e.,
   it removes Antrea)
3) some workload Pods are restarted / scheduled on the Node
4) the antrea-agent Pod is restarted late, and fixes the CNI conf file

We make the following changes:

1) we ensure that the antrea-eks-node-init script runs again in case of
   Node restart.
2) we change the execution flow of the script: we first wait for
   aws-node to start, then wait 10s to give it enough time to overwrite
   the CNI conf file, then wait for Antrea CNI installation, and finally
   we restart all containers.

We also update antrea-aks-node-init.yml and antrea-gke-node-init.yml to
use the same container image as antrea-eks-node-init.yml. Using v2
ensures that the script is run again if it is modified at runtime.

Signed-off-by: Antonin Bas <abas@vmware.com>