Fix clickhouse client race during batch commit
This change fixes an unsafe deque access in flow aggregator clickhouse
client. Reads of the deque including Len() and At() are not protected
with mutex during batchCommitAll(), which could run concurrently with
another routine writing items to deque with CacheSet().

To avoid holding the mutex for extended period of time, the following
behavior is adopted:
1. (With mutex) Pop front #Len() items off deque and store in buffer
2. (Without mutex) Transform records in buffer and batch export them
   with SQL driver. This step is where I/O happens.
3. (With mutex) If SQL commit rolled back, PushFront records in buffer
   back to deque. Oldest records will be discarded if deque full.

Also bumped github.com/gammazero/deque to v0.1.2.

Signed-off-by: Shawn Wang <wshaoquan@vmware.com>