Refactor FlowAggregator exporter code
We currently support 2 different exporters for the FlowAggregator: the
IPFIX exporter and the ClickHouse exporter. As I plan to add a third
one, which will upload flow record to AWS S3, it is good to unify
exporter implementations behind a single interface.

Before this change the IPFIX exporter was deeply embedded within the
main FlowAggregator logic. With this change, it is treated as any other
exporter.

This change also tries to improve thread-safety (in particular for live
updates of exporter configuration options) and adds a few unit tests.

Signed-off-by: Antonin Bas <abas@vmware.com>