[CI] Run integration tests with Github Action
Integration tests consume few resources and has no special requirements
on testbeds. This patch changes to run them with Github Action to reduce
dependencies on Jenkins and VM testbeds.

Besides, it fixes an issue in the test-vmc.sh that integration test
failures were ignored because two tests were connected with the
semi-colon operator.

Signed-off-by: Quan Tian <qtian@vmware.com>