Revert some of the changes to IPFIX exporter
At the moment (because of the go-ipfix implementation), it is not
possible to make the IPFIX exporter asynchronous like the ClickHouse
one, with a queue to buffer records that have yet to be sent. This is
because the Record object is still owned by the aggregation process, and
cannot be safely buffered.

This doesn't impact the main contribution of the PR, which is still to
abstract away the different exporters behind a common
interface. However, when AddRecord is called for the IPFIX exporter, the
recod is sent synchronously, just like it was before this PR.

Signed-off-by: Antonin Bas <abas@vmware.com>