[ExternalNode] Fix Agent log issue
Antrea Agent uses TrafficEncryptionModeInvalid in networkConfig. It may
introduce a panic when calling TrafficEncapModeInvalid.String() since
its string value is not defined in the string slice for the supported
TrafficEncryptionModeInvalid, and its value is -1 which is not a valid
index in any slice.

To resolve the issue, return a constant string when
TrafficEncapModeInvalid is used.

The same issue may exist on type TrafficEncapModeInvalid in the logging.
As a result, a defualt string value is given to TrafficEncapModeInvalid
too.

Signed-off-by: wenyingd <wenyingd@vmware.com>