Fix race condition when stopping Flow Aggregator
When watchConfiguration was stopped "before" flowExportLoop, it was
possible for flowExportLoop to read from a closed updateCh (channel used
for configuration updates). A nil value was returned and passed to
updateFlowAggregator, causing a panic. Such a panic was observed in CI
(TestFlowAggregator_Run) unit test.

To avoid this issue, we test if the channel is closed when reding from
it. If it is closed, we stop reading from it altogether.

Signed-off-by: Antonin Bas <abas@vmware.com>