Support updating the antrea/base-windows image with Github Actions (#4109)
At the moment, updating this image is a very manual process and requires
access to a Windows machine. To simplify this process, we add a new
Github Workflow which can be used to manually trigger a Docker image
build and push on Github hosted Windows workers.

Ideally, the workflow would not need to be triggered manually and would
run as part of CI (like we do for Linux base images). However this is a
good compromise / first step. Even with Docker caching, running
./hack/build-antrea-windows-all.sh for every pull request and push event
may be a bit too expensive. Something that can be investigated later.

Signed-off-by: Antonin Bas <abas@vmware.com>