Add checksum of config file as an annotation
When the ConfigMap is generated with suffix by kustomize, it will
generate a manifest with new ConfigMap like 'antrea-mc-controller-config-***'.
When a new Deployment is applied, it will leave a stale ConfigMap.
To avoid stale ConfigMaps, here we add checksum of multi-cluster configuration
file as Deployment's annotation. This way can help to restart the controller Pod
automatically without the suffix of ConfigMap.

Signed-off-by: Lan Luo <luola@vmware.com>