Fix log spam when there is any DNS based LoadBalancer Service
IP in Service.Status.LoadBalancer.Ingress is optional and may be not set
for load-balancer ingress points that are DNS based. It should check if
the IP is empty before appending it to the slice, otherwise every round
of sync will produce one error log "Skipping invalid IP:" for each IP,
leading to log spam.

The same issue has been fixed in K8s upstream by
https://github.com/kubernetes/kubernetes/pull/101690.

Signed-off-by: Quan Tian <qtian@vmware.com>