Use the docker.io registry for "latest" antrea images
It has been observed that pulling
projects.registry.vmware.com/antrea/antrea-ubuntu:latest will often
fail. This is because the Harbor registry uses a CDN for caching, which
doesn't work well at all when tags are being re-used (in our case, we
re-use the latest tag all the time). While most users should be
installing a released version of antrea (with an immutable tag) and thus
should not suffer from this issue, it can be frustrating for users who
want to install the latest antrea built from the main branch.

As a result, we switch to the docker.io registry whenever the latest tag
is used, and in particular for checked-in manifest files. The Harbor
registry (projects.registry.vmware.com) is still used for all release
assets, and more generally for all images with an immutable tag.

We are able to simplify some infrastructure code because of that, as
built images no longer need to be tagged with
projects.registry.vmware.com.

Signed-off-by: Antonin Bas <abas@vmware.com>