Remove redundant flow in IngressSecurityClassifierTable
In IngressSecurityClassifierTable, flows for forwarding the packets destined
for Antrea gateway, tunnel and uplink to IngressMetricTable are installed
(to bypass IngressMetricTable). The removed flow is used to enforce the +new+trk
(only +new+trk packet has ToNodePortAddressRegMark) NodePort to AntreaPolicyIngressRuleTable.
However, if the packet needs to be forwarded to the selected Endpoint through
Antrea gateway, tunnel and uplink, in another word, the selected Endpoint is not
a local Pod, as a result, it is not necessary to enforce the packet to
AntreaPolicyIngressRuleTable. If the Endpoint is a local Pod, the packet can be
forwarded AntreaPolicyIngressRuleTable since its destination is not Antrea gateway,
tunnel or uplink.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>