Label Identity replication for multi-cluster NetworkPolicies (#3664)
This commit is the first part of the multi-cluster stretched network
policy implementation. In order to enforce policies globally based on
label selections, Antrea MC Service needs to inform each member cluster
of all Pod + Namespace label combinations in the ClusterSet. Each
unique Pod + Namespace label combination is defined as labelidentity,
which is a normalized string calculated based on Pod's own labels and
its Namespace labels. A couple of new controllers and CRDs are added,
namely:
- A label_identity_controller in each member cluster, which watches Pod
and Namespace events, and updates ResourceExport of type labelidentity
for each unique label identity in the cluster.
- A label_identity_export_controller in the leader cluster, which
watches for ResourceExports of type labelidentity from all member
clusters, dedups the label identities and assigns a unique ID for each
for them. It also creates an ResourceImport of type labelidentity
object for the combination of each such labelidentity and its assigned
ID in the leader cluster.
- A label_identity_export_controller in each member cluster, which
watches for ResourceImport of type labelidentity events in the leader,
and creates a LabelIdentity CRD object in each member cluster for each
ResourceImport of type labelidentity.

Signed-off-by: Dyanngg <dingyang@vmware.com>
Co-authored-by: wgrayson <wgrayson@vmware.com>