Make tunnel csum option configurable and default to false (#4250)
For Linux kernel before Mar 2021, UDP checksum must be present to
trigger GRO on the receiver for better performance of Geneve and VXLAN
tunnels. The issue has been fixed in Linux kernel [1], thus computing
UDP checksum is no longer necessary.

This patch exposes a configuration parameter TunnelCsum which
determines whether to compute UDP encapsulation header (Geneve or
VXLAN) checksums on outgoing packets and makes it default to false. It
should only be set to true when Nodes run an unpatched Linux kernel and
poor transfer performance is observed.

[1] https://github.com/torvalds/linux/commit/89e5c58fc1e2857ccdaae506fb8bc5fed57ee063

Signed-off-by: Quan Tian <qtian@vmware.com>