Reset pipelines after pipelines are generated in every unit test
When generating pipelines, every table ID is allocated by calling
`binding.NextTableID()`. Note that, every time the function is called,
its return value (uint8) is increased by 1. After each unit test,
if not calling `binding.ResetTableID()` to reset the return value,
there is a chance that the return value overflows when function to
generate pipelines is called enough times. When overflowing, two
consecutive return values are 255 and 0, causing test TestBuildPipeline
to fail since 255 is not less than 0 (in TestBuildPipeline, there is
an asserting about the table of previous table should be less than
that of the current table).

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>