Enclose Pod labels with single quote in s3 uploader
In this commit, we updated the s3 uploader inside Flow Aggregator to
enclose Pod label fields with single quote for snowflake usecases.
Because commas are used to separate different columns in CSV and Pod
labels json string contains commas, we enclose it with single quote to
avoid them being separated into multiple columns in Snowflake.

Signed-off-by: Yongming Ding <dyongming@vmware.com>