Fix unit tests of TrafficControl
The tests should wait for podInformer to be synced, otherwise it's
indeterministic when the initial Pods will be processed, and caused flaky
tests:

In TestTrafficControlAdd, Pods may not be in cache when syncTrafficControl is
executed.

In TestPodLabelsUpdate, the initial Pod may be not in cache and the call that
updates the Pod may generate an ADD event, instead of an UPDATE event.

Besides, the patch uses a deterministic method to wait for an update has been
processed by informer to avoid unnecessary long time spent on sleeping.

Fixes #4197

Signed-off-by: Quan Tian <qtian@vmware.com>