Remove deadcode
The "deadcode" linter is deprecated (no longer maintained) and is now
replaced with "unused".
Running this new linter revealed some deadcode throughout the code base,
which is being removed by this PR.
I tried to make sure that in each instance, the code was actually
supposed to be removed, as opposed to indicative of a bug (e.g.,
function that should be called for correctness but that is not actually
called).

Signed-off-by: Antonin Bas <abas@vmware.com>