Fix base-windows Dockerfile and script
CNI_BINARIES_VERSION is missing when building windows-golang and
antrea/base-windows images, leading to CNI binaries missing in the
target image. The issue was not discovered because the curl command
used in the Dockerfile ignored HTTP 404 code and didn't fail the
build.

The patch adds the missing argument and makes curl command fail on
HTTP errors.

Signed-off-by: Quan Tian <qtian@vmware.com>