Extend Antrea native policy APIs to Layer 7
This patch extends Antrea native policy APIs to application layer. A
field named "L7Protocols" is added to the rule struct of Antrea
NetworkPolicy and ClusterNetworkPolicy APIs, representing set of Layer
7 protocols matched by the rule.

For now, if this field is set, action can only be Allow. When this field
is used in a rule, any traffic matching the other criteria of the rule
(typically the 5-tuple) will be forwarded to an application-aware
component for protocol detection and rule enforcement, and the traffic
will be allowed if the Layer 7 criteria is also matched, otherwise
dropped. Therefore, any rules after a Layer 7 rule will not be enforced
for this traffic.

For #4231

Signed-off-by: Quan Tian <qtian@vmware.com>